multi-db-user-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
asyncpg
sqlalchemy
pydantic[email]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart